Salt master no longer returns true when set to 'false'
The current default template just checks for the existence of the parameter, not its value. When an user sets the 'salt_master' parameter to 'false', it's expected to not enable Salt in the host, however it'll enable it. I changed it to check if the value of the parameter is the string 'true' as we do with 'force-puppet' above.